refactor: enum클래스 리패키징
