+++ src/main/java/lotto/exception/LottoGameException.java
@@ -1,5 +1,7 @@
 package lotto.exception;
 
+import lotto.enums.ErrorMessage;
+
 public class LottoGameException extends Exception {
     public LottoGameException(ErrorMessage message) {
         super(message.getMessage());
