+++ src/main/java/lotto/utils/LottoMachine.java
@@ -6,7 +6,7 @@ import lotto.exception.LottoGameException;
 
 import java.util.List;
 
-import static lotto.utils.UtilNumber.*;
+import static lotto.enums.UtilNumber.*;
 
 public class LottoMachine {
